refactor: Intent, IntentClassifier를 intent/ 패키지로 이동 (#96)
query ↔ retrieval 간 순환 의존을 제거하기 위해
Intent(enum)과 IntentClassifier를 별도 intent/ 패키지로 추출
