+++ src/main/java/com/flab/woowahaneats/domain/chatbot/application/ChatbotServiceImpl.java
@@ -3,8 +3,8 @@ package com.flab.woowahaneats.domain.chatbot.application;
 import com.flab.woowahaneats.domain.chatbot.controller.dto.ChatResponse;
 import com.flab.woowahaneats.domain.chatbot.retrieval.DataQuery;
 import com.flab.woowahaneats.domain.chatbot.retrieval.DataRetriever;
-import com.flab.woowahaneats.domain.chatbot.retrieval.Intent;
-import com.flab.woowahaneats.domain.chatbot.retrieval.IntentClassifier;
+import com.flab.woowahaneats.domain.chatbot.intent.Intent;
+import com.flab.woowahaneats.domain.chatbot.intent.IntentClassifier;
 import com.flab.woowahaneats.domain.chatbot.retrieval.RetrievalResult;
 import com.flab.woowahaneats.domain.restaurant.owner.service.OwnerRestaurantService;
 import lombok.RequiredArgsConstructor;
