+++ src/main/java/com/flab/woowahaneats/domain/chatbot/query/QueryTemplateRegistry.java
@@ -1,6 +1,6 @@
 package com.flab.woowahaneats.domain.chatbot.query;
 
-import com.flab.woowahaneats.domain.chatbot.retrieval.Intent;
+import com.flab.woowahaneats.domain.chatbot.intent.Intent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Component;
