+++ src/main/java/com/flab/woowahaneats/domain/chatbot/query/TemplateQueryExecutor.java
@@ -1,6 +1,6 @@
 package com.flab.woowahaneats.domain.chatbot.query;
 
-import com.flab.woowahaneats.domain.chatbot.retrieval.Intent;
+import com.flab.woowahaneats.domain.chatbot.intent.Intent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
