+++ src/main/java/com/flab/woowahaneats/domain/chatbot/retrieval/DataQuery.java
@@ -1,5 +1,7 @@
 package com.flab.woowahaneats.domain.chatbot.retrieval;
 
+import com.flab.woowahaneats.domain.chatbot.intent.Intent;
+
 public record DataQuery(
         Intent intent,
         String originalMessage,
